% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identical_vectorised.r
\name{aj_identical_vectorised}
\alias{aj_identical_vectorised}
\title{Identical Vectorised}
\usage{
aj_identical_vectorised(x, y)
}
\description{
this function checks to see if two columns are equal, row by row.
It handles cases where NA exists too
}
