% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_input_file.r
\name{aj_input_file}
\alias{aj_input_file}
\title{Get Filename}
\usage{
aj_input_file(
    filename,
    output_folder = "./",
    extension = "xlsx"
)
}
\arguments{
\item{filename}{The filename used (excluding version control and file extension).}

\item{output_folder}{The folder where the outputs should be saved.}

\item{extension}{I'm using .csv and/or .xlsx, but can be anything.}
}
\description{
Allows the user to define a filename with version control.
The latest version of said file is then identified and returned.
This is the sister function to \code{aj_output_file()} and looks for files using that naming convention.
}
\examples{
list.files(path = "./outputs/")
# [1] output1_v1.0.xlsx  output1_v1.1.xlsx  output1_v1.2.xlsx  output2_v1.0.xlsx

aj_input_file(
    filename = "output1",
    output_folder = "./outputs/",
    extension = "xlsx"
)
# [1] "./outputs/output1_v1.2.xlsx"
}
