% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_output_file.r
\name{aj_output_file}
\alias{aj_output_file}
\title{Set Filename}
\usage{
aj_output_file(
    filename,
    output_folder = "./",
    major_v = 0,
    extension = "xlsx"
)
}
\arguments{
\item{filename}{The filename used (excluding version control and file extension).}

\item{output_folder}{The folder where the outputs should be saved.}

\item{major_v}{The major version number.}

\item{extension}{I'm using .csv and/or .xlsx, but can be anything.}
}
\description{
Allows the user to define a filename with version control.
Minor version increments with each additional file with the defined filename and given major version number.
Note: incremental version control breaks with multiple files using the same name.
}
\examples{
list.files(path = "./outputs/")
# [1] output1_v1.0.xlsx  output1_v1.1.xlsx  output1_v1.2.xlsx  output2_v1.0.xlsx

aj_output_file(
    filename = "output1",
    output_folder = "./outputs/",
    major_v = 1,
    extension = "xlsx"
)
# [1] "./outputs/output1_v1.3.xlsx"

aj_output_file(
    filename = "output1",
    output_folder = "./outputs/",
    major_v = 2,
    extension = "xlsx"
)
# [1] "./outputs/output1_v2.0.xlsx"
}
