% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_round2.r
\name{aj_round2}
\alias{aj_round2}
\title{Rounding Function}
\usage{
aj_round2(x, digits = 0)
}
\arguments{
\item{x}{The number to round}

\item{digits}{The number of digits to round to.}
}
\description{
Base R Rounding:
Note that for rounding off a 5, the IEC 60559 standard (see also ‘IEEE 754’) is expected to be used, ‘go to the even
digit’. Therefore round(0.5) is 0 and round(-1.5) is -2. However, this is dependent on OS services and on
representation error (since e.g. 0.15 is not represented exactly, the rounding rule applies to the represented number
and not to the printed number, and so round(0.15, 1) could be either 0.1 or 0.2).

Issue:
This method of rounding is not consistent with rounding from other sources (e.g. Excel) and not inline with our
needs.

Solution:
This function gets around this by identifying the specific digit to be rounded by, then applying "ceiling" or
"floor" functions to round up or down as necessary.
This function has been designed to run with both single digits and digit vectors.
}
\examples{
aj_round2(3.141592654, digits = 3)
# [1] 3.142

aj_round2(314.1592654, digits = -1)
# [1] 310
}
